@@ -17,57 +17,21 @@ abstract class AbstractConverter
 {
     use LoggerAwareTrait;
 
-    protected $baseSchemas = [
+    protected array $baseSchemas = [
         'http://www.w3.org/2001/XMLSchema',
         'http://www.w3.org/XML/1998/namespace',
     ];
 
-    protected $namespaces = [
+    protected array $namespaces = [
         'http://www.w3.org/2001/XMLSchema' => '',
         'http://www.w3.org/XML/1998/namespace' => '',
     ];
 
-    /**
-     * @var \GoetasWebservices\Xsd\XsdToPhp\Naming\NamingStrategy
-     */
-    private $namingStrategy;
+    private NamingStrategy $namingStrategy;
 
-    abstract public function convert(array $schemas);
+    protected array $typeAliases = [];
 
-    protected $typeAliases = [];
-
-    protected $aliasCache = [];
-
-    public function addAliasMap($ns, $name, callable $handler)
-    {
-        $this->logger->info("Added map $ns $name");
-        $this->typeAliases[$ns][$name] = $handler;
-    }
-
-    public function addAliasMapType($ns, $name, $type)
-    {
-        $this->addAliasMap($ns, $name, function () use ($type) {
-            return $type;
-        });
-    }
-
-    public function getTypeAliases(): array
-    {
-        return $this->typeAliases;
-    }
-
-    public function getTypeAlias($type, ?Schema $schemapos = null)
-    {
-        $schema = $schemapos ?: $type->getSchema();
-
-        $cid = $schema->getTargetNamespace() . '|' . $type->getName();
-        if (isset($this->aliasCache[$cid])) {
-            return $this->aliasCache[$cid];
-        }
-        if (isset($this->typeAliases[$schema->getTargetNamespace()][$type->getName()])) {
-            return $this->aliasCache[$cid] = call_user_func($this->typeAliases[$schema->getTargetNamespace()][$type->getName()], $type);
-        }
-    }
+    protected array $aliasCache = [];
 
     public function __construct(NamingStrategy $namingStrategy, ?LoggerInterface $logger = null)
     {
@@ -186,15 +150,48 @@ abstract class AbstractConverter
         });
     }
 
-    /**
-     * @return \GoetasWebservices\Xsd\XsdToPhp\Naming\NamingStrategy
-     */
-    protected function getNamingStrategy()
+    abstract public function convert(array $schemas);
+
+    public function addAliasMap(string $ns, string $name, callable $handler): void
+    {
+        $this->logger->info("Added map $ns $name");
+        $this->typeAliases[$ns][$name] = $handler;
+    }
+
+    public function addAliasMapType(string $ns, string $name, $type): void
+    {
+        $this->addAliasMap($ns, $name, function () use ($type) {
+            return $type;
+        });
+    }
+
+    public function getTypeAliases(): array
+    {
+        return $this->typeAliases;
+    }
+
+    public function getTypeAlias($type, Schema $schemapos = null)
+    {
+        $schema = $schemapos ?: $type->getSchema();
+
+        $cid = $schema->getTargetNamespace() . '|' . $type->getName();
+        if (isset($this->aliasCache[$cid])) {
+            return $this->aliasCache[$cid];
+        }
+        if (isset($this->typeAliases[$schema->getTargetNamespace()][$type->getName()])) {
+            return $this->aliasCache[$cid] =
+                call_user_func($this->typeAliases[$schema->getTargetNamespace()][$type->getName()], $type);
+        }
+
+        return null;
+    }
+
+    protected function getNamingStrategy(): NamingStrategy
     {
         return $this->namingStrategy;
     }
 
-    public function addNamespace($ns, $phpNamespace)
+    public function addNamespace(string $ns, string $phpNamespace): static
     {
         $this->logger->info("Added ns mapping $ns, $phpNamespace");
         $this->namespaces[$ns] = $phpNamespace;
@@ -202,15 +199,12 @@ abstract class AbstractConverter
         return $this;
     }
 
-    protected function cleanName($name)
+    protected function cleanName(string $name): string
     {
         return preg_replace('/<.*>/', '', $name);
     }
 
-    /**
-     * @return \GoetasWebservices\XML\XSDReader\Schema\Type\Type|null
-     */
-    protected function isArrayType(Type $type)
+    protected function isArrayType(Type $type): ?Type
     {
         if ($type instanceof SimpleType) {
             if ($type->getList()) {
@@ -227,31 +221,27 @@ abstract class AbstractConverter
         return  null;
     }
 
-    /**
-     * @return \GoetasWebservices\XML\XSDReader\Schema\Element\ElementSingle|null
-     */
-    protected function isArrayNestedElement(Type $type)
+    protected function isArrayNestedElement(Type $type): ?ElementSingle
     {
         if ($type instanceof ComplexType && !$type->getParent() && !$type->getAttributes() && count($type->getElements()) === 1) {
             $elements = $type->getElements();
 
             return $this->isArrayElement(reset($elements));
         }
+
+        return null;
     }
 
-    /**
-     * @param mixed $element
-     *
-     * @return \GoetasWebservices\XML\XSDReader\Schema\Element\ElementSingle|null
-     */
-    protected function isArrayElement($element)
+    protected function isArrayElement(mixed $element): ?ElementSingle
     {
         if ($element instanceof ElementSingle && ($element->getMax() > 1 || $element->getMax() === -1)) {
             return $element;
         }
+
+        return null;
     }
 
-    public function getNamespaces()
+    public function getNamespaces(): array
     {
         return $this->namespaces;
     }
